requirements.txt
# Market Data & Technical Analysis
yfinance>=0.2.33
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# NLP & Sentiment Analysis
transformers>=4.35.0
torch>=2.0.0
feedparser>=6.0.10

# Deep Learning
tensorflow>=2.14.0
keras>=2.14.0
scikit-learn>=1.3.0

# Data Visualization & Dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
schedule>=1.2.0
psutil>=5.9.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
